Move delete workflow button to top nav bar
